+++ spec/lightstreamer/protocol_error_spec.rb
@@ -3,7 +3,7 @@ describe Lightstreamer::ProtocolError do
     error = Lightstreamer::ProtocolError.new 'message', '15'
 
     expect(error.error).to eq('message')
-    expect(error.code).to be_nil
+    expect(error.code).to eq(15)
     expect(error.message).to eq('Lightstreamer error: message, code: 15')
   end
 end