+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 0.2 — Unreleased
 
+- Added `Lightstreamer::Session#disconnect` and `Lightstreamer::StreamConnection#disconnect`
+- Fixed `Lightstreamer::ProtocolError#code` not being set
 - Renamed the `--address` option to the command-line client to `--server-url`
 
 ### 0.1 — July 22, 2016